requirements.txt
Django==3.0.8
djangorestframework==3.11.0
PyMeeus==0.3.7